Relocate model_assessment_to_verifier_repair_assessment to verifier_orchestration
Phase 2 batch 7 for Issue #682. Moves the 166-LOC legacy boundary
helper out of turn.rs into verifier_orchestration.rs.

Moved (turn.rs -> verifier_orchestration.rs):

  - model_assessment_to_verifier_repair_assessment (166 LOC) —
    SF3 legacy boundary helper that converts a parsed diagnostic
    reply into a `VerifierRepairAssessment` value. Orchestrates the
    full repair-candidate pipeline (recovery_target_hint_for_diagnostic_path
    + diagnostic_target_allowed_by_confidence + admit_repair_target_hint
    + verifier_repair_missing_local_module_provider /
    preferred_local_import_source / stale_assertion_test_target +
    first_role_kind_compatible_diagnostic_target) and assembles the
    final assessment with the 7-field struct literal.

The fn is promoted to pub(super) (file-private before). The SF3 doc
comment (Responsibility boundary / Legacy boundary vs Semantic
boundary contract) is preserved verbatim above the relocated
function in the new module.

Adjustments:

  * verifier_orchestration.rs: added imports for
    - super::VerifierRepairAssessmentSource (parent loop_run module)
    - super::repair_target_admission::{RepairTargetAdmissionContext,
      admit_repair_target_hint}
    - super::semantic_repair_planning::{
      diagnostic_target_allowed_by_confidence,
      first_role_kind_compatible_diagnostic_target}
    - super::task_contract::ArtifactRole (the Setup-vs-other gate
      reads via failure_kind.allows_setup_target())
    - super::verifier_assessment_parser::{ParsedVerifierRepairAssessment,
      verifier_failure_type_for_diagnostic_kind}
    - super::verifier_repair_targeting::{
      recovery_target_hint_for_diagnostic_path,
      verifier_repair_missing_local_module_provider,
      verifier_repair_preferred_local_import_source,
      verifier_repair_stale_assertion_test_target}

  * turn.rs:
    - Extended use super::verifier_orchestration::{...} with
      model_assessment_to_verifier_repair_assessment.
    - Dropped admit_repair_target_hint /
      diagnostic_target_allowed_by_confidence /
      first_role_kind_compatible_diagnostic_target /
      ParsedVerifierRepairAssessment /
      verifier_failure_type_for_diagnostic_kind /
      recovery_target_hint_for_diagnostic_path /
      verifier_repair_missing_local_module_provider /
      verifier_repair_preferred_local_import_source /
      verifier_repair_stale_assertion_test_target from production
      imports.
    - Re-added admit_repair_target_hint /
      diagnostic_target_allowed_by_confidence /
      ParsedVerifierRepairAssessment /
      verifier_repair_preferred_local_import_source /
      verifier_repair_stale_assertion_test_target under #[cfg(test)]
      so the in-mod test fixtures that build literals + the
      mod progress_tests `use super::{...}` references continue to
      resolve.
    - Cleaned 2 orphaned doc comment blocks left over the moved
      function and the verifier_repair_decision wrapper (clippy
      empty_line_after_doc_comments + doc_list_item_without_indentation
      under -D warnings).
    - Updated `progress_tests::sf3_model_assessment_doc_comment_pins_legacy_boundary`
      to read the boundary fn from verifier_orchestration.rs with the
      new pub(super) marker and to count VerifierRepairAssessment
      literals across both files (preserves the boundary invariant
      across the module split).

Metrics: turn.rs 33,998 -> 33,796 (-202); verifier_orchestration.rs
1,072 -> 1,280 (+208). Cumulative Issue #682 progress:
turn.rs 34,958 -> 33,796 (-1,162, ~39% of Phase 2 target).
Acceptance gap: 33,796 -> 32,000 = ~1,796 LOC remaining.
